closet_app/requirements_gemini.txt
google-genai==1.34.0
gradio==5.44.1
pdfminer.six==20250506
ipykernel==6.30.1
ipywidgets==8.1.7
langchain==0.3.27
langchain-google-genai==2.1.10
langchain-community==0.3.29
faiss-cpu==1.12.0
pillow==11.3.0
